migrate useIgnoredIds to TS Query V5
