--- src/pages/CommitDetailPage/hooks/useIgnoredIds.ts
@@ -1,10 +1,0 @@
-import { useQuery } from '@tanstack/react-query'
-
-export const useIgnoredIds = () =>
-  useQuery<number[]>({
-    queryKey: ['IgnoredUploadIds'],
-    queryFn: () => Promise.resolve([]),
-    suspense: false,
-    staleTime: Infinity,
-    initialData: [],
-  })
